chore: expand fixed player panel to full width
In previous design, fixed players can expand their panels to 400px
width, thus leading to asymmetry especially on PC (which is often much
wider than 400px). Now the panel can be expanded to full width of the
webpage.

It uses a (not very) new function in CSS: `calc()`, which allows making
calculation with percentages (%) and pixels (px) without JavaScript.
Note that here uses `calc(100% - 18px)`, where 18px is a magic number -
it actually the width of "panel button" (controls whether the panel to
be expanded or folded). We don't want to introduce more variables, for
the width is used in some JavaScript code too, and the variables in the
two system can't be mixed naturally.

Another point, we can't use `100%` for width, because the width of the
panel button is just covered by right padding of player body. If we do
that, the button will overflow when the player expands.