feat: allow to add songs at any index of playlist
Now APlayer supports to song addition at any index of a playlist. In the
past it was only allowed to append songs at the end of the list.

To use this new feature, call `ap.list.add()` function with both `audios`
and `index`. `index` is set to `null` if nothing is passed to it, and it
acts like just appending to the end.