fix: player panel doesn't expand to full as a dependency
When this library behaves a dependency of another project, it seems that
the player panel still doesn't expand to full width of webpage. In dev
tool of browsers we found that one style source, `index.scss` from
webpack, doesn't reset `max-width` property to `initial` (to override
values from other sources). So we guess we should explicitly set its
value and check whether the bug is fixed or not.